requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
scikit-learn
numpy
requests
python-dotenv
groq
